playground/ag_011_Event_Handler/pyproject.toml
[project]
name = "ag-011-event-handler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.193.0",
    "google-auth-httplib2>=0.3.1",
    "google-auth-oauthlib>=1.3.1",
    "huggingface-hub>=1.9.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]
